Refactor filenames to kebab case to follow aurelia convention.
closes #1
