+++ src/app.html
@@ -1,6 +1,6 @@
 <template>
-  <require from="./shared/layouts/headerlayout"></require>
-  <require from="./shared/layouts/footerlayout"></require>
+  <require from="./shared/layouts/header-layout"></require>
+  <require from="./shared/layouts/footer-layout"></require>
 
   <header-layout router-config.bind="router.currentInstruction.config"></header-layout>
   <router-view></router-view>
