+++ src/components/auth/auth-component.html
@@ -1,6 +1,4 @@
 <template>
-  <require from="../../resources/value-converters/keys"></require>
-
   <div class="auth-page">
     <div class="container page">
       <div class="row">
