+++ src/components/editor/editor-component.js
@@ -1,12 +1,12 @@
 import {inject, observable} from 'aurelia-framework';
 import {Router} from 'aurelia-router'
-import {ArticleService} from "../../shared/services/articleservice";
+import {ArticleService} from "../../shared/services/article-service";
 
 @inject(ArticleService, Router)
 export class EditorComponent {
   article;
   @observable() tag;
-  
+
   constructor(as, r) {
     this.articleService = as;
     this.router = r;
@@ -31,20 +31,20 @@ export class EditorComponent {
     }
     return null;
 }
-  
+
   tagChanged(newValue, oldValue) {
     if (newValue !== undefined && newValue !== '')
       this.addTag(this.tag);
   }
-  
+
   addTag(tag) {
     this.article.tagList.push(tag);
   }
-  
+
   removeTag(tag) {
     this.article.tagList.splice(this.article.tagList.indexOf(tag), 1);
   }
-  
+
   publishArticle() {
     this.articleService.save(this.article)
       .then((article) => {
