+++ src/components/home/home-component.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../../resources/elements/articlelist"></require>
+  <require from="../../resources/elements/article-list"></require>
 
   <div class="home-page">
