+++ src/components/home/home-component.js
@@ -1,8 +1,8 @@
 import {BindingEngine} from 'aurelia-framework';
 import {inject} from 'aurelia-dependency-injection';
-import {SharedState} from '../../shared/state/sharedstate';
-import {ArticleService} from "../../shared/services/articleservice"
-import {TagService} from '../../shared/services/tagservice';
+import {SharedState} from '../../shared/state/shared-state';
+import {ArticleService} from "../../shared/services/article-service"
+import {TagService} from '../../shared/services/tag-service';
 
 @inject(SharedState, BindingEngine, ArticleService, TagService)
 export class HomeComponent {
@@ -14,30 +14,30 @@ export class HomeComponent {
   totalPages;
   currentPage = 1;
   limit = 10;
-  
+
   constructor(sharedState, bindingEngine, articleService, tagService) {
     this.sharedState = sharedState;
     this.bindingEngine = bindingEngine;
     this.articleService = articleService;
     this.tagService = tagService;
   }
-  
+
   bind() {
     this.subscription = this.bindingEngine.propertyObserver(this.sharedState, 'isAuthenticated')
       .subscribe((newValue, oldValue) => {
         console.log('homeComponent isAuthenticated: ', newValue)
       })
   }
-  
+
   unbind() {
     this.subscription.dispose();
   }
-  
+
   attached() {
     this.getArticles();
     this.getTags();
   }
-  
+
   getArticles() {
     let params = {
       limit: this.limit,
@@ -49,12 +49,12 @@ export class HomeComponent {
       .then(response => {
         this.articles.splice(0);
         this.articles.push(...response.articles)
-  
+
         // Used from http://www.jstips.co/en/create-range-0...n-easily-using-one-line/
         this.totalPages = Array.from(new Array(Math.ceil(response.articlesCount / this.limit)), (val, index) => index + 1);
       })
   }
-  
+
   getTags() {
     this.tagService.getList()
       .then(response => {
@@ -62,14 +62,14 @@ export class HomeComponent {
         this.tags.push(...response);
       })
   }
-  
+
   setListTo(type, tag) {
     if (type === 'feed' && !this.sharedState.isAuthenticated) return;
     this.shownList = type;
     this.filterTag = tag;
     this.getArticles();
   }
-  
+
   getFeedLinkClass() {
     let clazz = '';
     if (!this.sharedState.isAuthenticated)
@@ -78,7 +78,7 @@ export class HomeComponent {
       clazz += ' active';
     return clazz;
   }
-  
+
   setPageTo(pageNumber) {
     this.currentPage = pageNumber;
     this.getArticles();
