+++ src/components/profile/profile-article-component.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../../resources/elements/articlelist"></require>
+  <require from="../../resources/elements/article-list"></require>
 
   <article-list articles.bind="articles" total-pages.bind="totalPages" page-number.bind="pageNumber"
                 current-page.bind="currentPage" set-page-cb.call="setPageTo(pageNumber)"></article-list>
