+++ src/resources/elements/article-list.js
