+++ src/resources/elements/article-preview.html
