+++ src/resources/elements/article-preview.js
