+++ src/resources/index.js
@@ -1,3 +1,7 @@
 export function configure(config) {
-  config.globalResources(['./value-converters/date']);
+  config.globalResources([
+    './value-converters/date',
+    './value-converters/format-html',
+    './value-converters/keys',
+  ]);
 }
