+++ src/shared/services/article-service.js
@@ -1,5 +1,5 @@
 import {inject} from 'aurelia-dependency-injection';
-import {ApiService} from './apiservice';
+import {ApiService} from './api-service';
 
 @inject(ApiService)
 export class ArticleService {
@@ -7,20 +7,20 @@ export class ArticleService {
   constructor(apiService) {
     this.apiService = apiService;
   }
-  
+
   getList(type, params) {
     return this.apiService.get('/articles' + ((type === 'feed') ? '/feed' : ''), params)
   }
-  
+
   get(slug) {
     return this.apiService.get('/articles/' + slug)
       .then(data => data.article)
   }
-  
+
   destroy(slug) {
     return this.apiService.delete('/articles/' + slug)
   }
-  
+
   save(article) {
     if (article.slug) {
       // If we're updating an existing article
@@ -32,13 +32,13 @@ export class ArticleService {
         .then(data => data.article)
     }
   }
-  
+
   favorite(slug) {
     return this.apiService.post('/articles/' + slug + '/favorite')
   }
-  
+
   unfavorite(slug) {
     return this.apiService.delete('/articles/' + slug + '/favorite')
   }
-  
+
 }
