+++ src/shared/services/tag-service.js
@@ -1,5 +1,5 @@
 import {inject} from 'aurelia-dependency-injection';
-import {ApiService} from './apiservice';
+import {ApiService} from './api-service';
 
 @inject(ApiService)
 export class TagService {
@@ -7,7 +7,7 @@ export class TagService {
   constructor(apiService) {
     this.apiService = apiService;
   }
-  
+
   getList() {
     return this.apiService.get('/tags')
       .then(data => data.tags)
