+++ src/shared/services/user-service.js
@@ -1,18 +1,18 @@
 import {inject} from 'aurelia-dependency-injection';
-import {ApiService} from './apiservice';
-import {JwtService} from './jwtservice';
+import {ApiService} from './api-service';
+import {JwtService} from './jwt-service';
 import {User} from '../models/user';
-import {SharedState} from '../state/sharedstate';
+import {SharedState} from '../state/shared-state';
 
 @inject(ApiService, JwtService, SharedState)
 export class UserService {
-  
+
   constructor(apiService, jwtService, sharedState) {
     this.apiService = apiService;
     this.jwtService = jwtService;
     this.sharedState = sharedState;
   }
-  
+
   // Verify JWT in localstorage with server & load user's info.
   // This runs once on application startup.
   populate() {
@@ -24,21 +24,21 @@ export class UserService {
       this.purgeAuth();
     }
   }
-  
+
   setAuth(user) {
     // Save JWT sent from server in localstorage
     this.jwtService.saveToken(user.token);
     this.sharedState.currentUser = user;
     this.sharedState.isAuthenticated = true;
   }
-  
+
   purgeAuth() {
     // Remove JWT from localstorage
     this.jwtService.destroyToken();
     this.sharedState.currentUser = new User();
     this.sharedState.isAuthenticated = false;
   }
-  
+
   attemptAuth(type, credentials) {
     const route = (type === 'login') ? '/login' : '';
     return this.apiService.post('/users' + route, {user: credentials})
@@ -47,13 +47,13 @@ export class UserService {
         return data;
       });
   }
-  
+
   update(user) {
     return this.apiService.put('/user', { user })
       .then(data => {
         this.sharedState.currentUser = data.user;
         return data.user;
       });
-    
+
   }
 }
