+++ src/shared/state/shared-state.js
@@ -1,7 +1,7 @@
 import {User} from '../models/user';
 
 export class SharedState {
-  
+
   constructor() {
     this.currentUser = new User();
     this.isAuthenticated = false;
